[로또] step1: feedback #2
