+++ src/main/java/step1/calculator/README.md
@@ -15,6 +15,6 @@
 ## # 피드백
 
 - [x] #1. k요구사항을 작성한다.
-- [ ] #2. 검증 로직과 비즈니스 로직을 분리한다.
+- [x] #2. 검증 로직과 비즈니스 로직을 분리한다.
 - [ ] #3. `switch/case` 대신 `enum` 을 활용한다.
 - [ ] #4. "CalculatorApp" 의 책임을 이관하고 세분화한다.
