+++ src/main/java/step1/calculator/controller/CalculatorApp.java
@@ -5,6 +5,7 @@ import step1.calculator.model.AddModule;
 import step1.calculator.model.DivideModule;
 import step1.calculator.model.MultiplyModule;
 import step1.calculator.model.SubtractModule;
+import step1.calculator.model.Validator;
 import step1.calculator.view.InputView;
 import step1.calculator.view.OutputView;
 
@@ -12,6 +13,7 @@ public class CalculatorApp {
 
     private static final InputView inputView = new InputView();
     private static final OutputView outputView = new OutputView();
+    private static final Validator validator = new Validator();
 
     private static final AddModule addModule = new AddModule();
     private static final SubtractModule subtractModule = new SubtractModule();
@@ -20,55 +22,17 @@ public class CalculatorApp {
 
     public void run() {
         String input = inputView.getInput();
+        validator.validate(input);
         int result = calculates(input);
         outputView.print(result);
     }
 
-    private void validate(String input) {
-        if (input == null || input.isBlank()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validate(String[] tokens) {
-        if (tokens == null || tokens.length % 2 == 0) {
-            throw new IllegalArgumentException();
-        }
-        IntStream.range(0, tokens.length).filter(i -> i % 2 == 0).filter(i -> !isNumber(tokens[i])).findFirst()
-            .ifPresent(i -> {
-                throw new IllegalArgumentException();
-            });
-        IntStream.range(0, tokens.length).filter(i -> i % 2 == 1).filter(i -> !isOperator(tokens[i])).findFirst()
-            .ifPresent(i -> {
-                throw new IllegalArgumentException();
-            });
-    }
-
-    private boolean isNumber(String token) {
-        try {
-            Integer.parseInt(token);
-            return true;
-        } catch (NumberFormatException e) {
-            return false;
-        }
-    }
-
-    private boolean isOperator(String token) {
-        return token.equals("+") || token.equals("-") || token.equals("*") || token.equals("/");
-    }
-
     public int calculates(String input) {
-        validate(input);
         String[] tokens = input.split(" ");
-        validate(tokens);
         return calculates(tokens);
     }
 
     public int calculates(String[] tokens) {
-        if (tokens == null || tokens.length == 0) {
-            throw new IllegalArgumentException();
-        }
-
         int result = Integer.parseInt(tokens[0]);
         for (int i = 1; i < tokens.length; i += 2) {
             result = calculate(result, tokens[i], Integer.parseInt(tokens[i + 1]));
